requirements.txt
# PiCar-X Hardware Control
RPi.GPIO>=0.7.1
adafruit-circuitpython-servokit>=1.3.7

# Camera
picamera2>=0.3.12

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0

# Coral USB Accelerator / TensorFlow Lite
pycoral>=2.0.0
tflite-runtime>=2.14.0

# Optional: For development and testing
pytest>=7.4.0
